+++ vss_cli/plugins/compute_plugins/vm.py
@@ -769,6 +769,64 @@ def compute_vm_get_template(ctx: Configuration):
     ctx.echo(format_output(ctx, [obj], columns=columns, single=True))
 
 
+@compute_vm_get.command('gpu', short_help='GPU configuration')
+@pass_context
+def compute_vm_get_gpu(ctx: Configuration):
+    """Virtual machine gpu configuration."""
+    columns = ctx.columns or const.COLUMNS_VM_GPU
+    objs = ctx.get_vm_gpu(ctx.moref)
+    ctx.echo(format_output(ctx, objs, columns=columns))
+
+
+@compute_vm_get.group('restore-point', short_help='Restore Points')
+@pass_context
+def compute_vm_get_rp(ctx: Configuration):
+    """Virtual machine restore points configuration."""
+
+
+@compute_vm_get_rp.command(
+    'ls', short_help='list virtual machine restore points'
+)
+@so.filter_opt
+@so.all_opt
+@so.page_opt
+@so.sort_opt
+@so.count_opt
+@pass_context
+def compute_vm_get_rp_ls(
+    ctx: Configuration,
+    filter_by,
+    show_all: bool,
+    sort,
+    page,
+    count,
+):
+    """List virtual machine restore points."""
+    params = dict(expand=1, sort='name,asc')
+    if all(filter_by):
+        params['filter'] = ';'.join(filter_by)
+    if all(sort):
+        params['sort'] = ';'.join(sort)
+    with ctx.spinner(disable=ctx.debug):
+        objs = (
+            ctx._get_objects(
+                f'/vm/{ctx.moref}/rp',
+                show_all=show_all,
+                per_page=count,
+                **params,
+            )
+            or []
+        )
+    columns = ctx.columns or const.COLUMNS_VM_RESTORE_POINTS
+    # format output
+    output = format_output(ctx, objs, columns=columns)
+    # page
+    if page:
+        click.echo_via_pager(output)
+    else:
+        ctx.echo(output)
+
+
 @compute_vm_get.command('tpm', short_help='vTPM configuration')
 @pass_context
 def compute_vm_get_tpm(ctx: Configuration):
@@ -1203,30 +1261,44 @@ def compute_vm_set_client(ctx: Configuration, client):
 
 
 @compute_vm_set.command('client-note', short_help='Client note (Metadata)')
-@click.argument('notes', required=True)
+@click.argument('notes', required=False)
+@click.option(
+    '--action',
+    type=click.Choice(['up', 'del']),
+    help='Action to perform.',
+    default='add',
+)
 @click.option(
     '--replace',
     '-r',
     is_flag=True,
     required=False,
-    help="Whether to replace existing value.",
+    help="Whether to replace existing value. Only works with up.",
 )
 @pass_context
-def compute_vm_set_client_note(ctx: Configuration, notes, replace):
+def compute_vm_set_client_note(ctx: Configuration, notes, replace, action):
     """Set or update virtual machine client notes in metadata.
 
-    vss-cli compute vm set <name-or-vm_id> client-note 'New note'
+    vss-cli compute vm set <name-or-vm_id> client-note --up 'New note'
     """
-    if not replace:
-        _old_notes = ctx.get_vm_notes(ctx.moref)
-        old_notes = _old_notes.get('value') or ""
-        notes = f"{old_notes}\n{notes}"
-    # generate payload
-    payload = dict(vm_id=ctx.moref, notes=notes)
-    # add common options
-    payload.update(ctx.payload_options)
     # request
-    obj = ctx.update_vm_notes(**payload)
+    if action == 'del':
+        # TODO: implement ctx.del_vm_notes from pyvss.
+        objs = ctx.request(f'/vm/{ctx.moref}/note/client', method=ctx.DELETE)
+        if objs:
+            obj = objs.get('data')
+    else:
+        if not notes:
+            raise click.BadArgumentUsage('Missing notes argument')
+        if not replace:
+            _old_notes = ctx.get_vm_notes(ctx.moref)
+            old_notes = _old_notes.get('value') or ""
+            notes = f"{old_notes}\n{notes}"
+        # generate payload
+        payload = dict(vm_id=ctx.moref, notes=notes)
+        # add common options
+        payload.update(ctx.payload_options)
+        obj = ctx.update_vm_notes(**payload)
     # print
     columns = ctx.columns or const.COLUMNS_REQUEST_SUBMITTED
     ctx.echo(format_output(ctx, [obj], columns=columns, single=True))